Move RemoveFile out of utils:files
Bug: b/543573091
