+++ base/cvd/cuttlefish/host/commands/assemble_cvd/BUILD.bazel
@@ -84,6 +84,7 @@ cf_cc_binary(
         "//cuttlefish/host/libs/feature:inject",
         "//cuttlefish/host/libs/log_names",
         "//cuttlefish/io:string",
+        "//cuttlefish/posix:remove",
         "//cuttlefish/posix:symlink",
         "//cuttlefish/pretty:vector",
         "//cuttlefish/result",
@@ -111,6 +112,7 @@ cf_cc_library(
         "//cuttlefish/host/libs/config:kernel_args",
         "//cuttlefish/host/libs/config:mkenvimage_slim",
         "//cuttlefish/host/libs/vm_manager",
+        "//cuttlefish/posix:remove",
         "//cuttlefish/result",
         "@abseil-cpp//absl/log",
         "@abseil-cpp//absl/strings",
@@ -144,6 +146,7 @@ cf_cc_library(
         "//cuttlefish/io:shared_fd",
         "//cuttlefish/io:string",
         "//cuttlefish/io:write_exact",
+        "//cuttlefish/posix:remove",
         "//cuttlefish/process:command",
         "//cuttlefish/process:execute",
         "//cuttlefish/result",
