+++ base/cvd/cuttlefish/host/commands/cvd/cli/BUILD.bazel
@@ -74,10 +74,10 @@ cf_cc_library(
     srcs = ["log_files.cpp"],
     hdrs = ["log_files.h"],
     deps = [
-        "//cuttlefish/common/libs/utils:files",
         "//cuttlefish/files:directory_contents",
         "//cuttlefish/files:directory_exists",
         "//cuttlefish/host/commands/cvd/utils",
+        "//cuttlefish/posix:remove",
         "//cuttlefish/result",
         "@abseil-cpp//absl/log",
         "@abseil-cpp//absl/strings",
@@ -167,6 +167,7 @@ cf_cc_library(
         "//cuttlefish/host/commands/cvd/instances/lock",
         "//cuttlefish/host/commands/cvd/utils",
         "//cuttlefish/posix:realpath",
+        "//cuttlefish/posix:remove",
         "//cuttlefish/posix:strerror",
         "//cuttlefish/posix:symlink",
         "//cuttlefish/result",
