+++ base/cvd/cuttlefish/host/commands/cvd/cli/commands/BUILD.bazel
@@ -387,6 +387,7 @@ cf_cc_library(
         "//cuttlefish/host/libs/log_names",
         "//cuttlefish/host/libs/metrics:device_metrics_orchestration",
         "//cuttlefish/io:string",
+        "//cuttlefish/posix:remove",
         "//cuttlefish/posix:symlink",
         "//cuttlefish/process:command",
         "//cuttlefish/result",
