+++ base/cvd/cuttlefish/host/commands/cvd/cli/commands/start.cpp
@@ -77,6 +77,7 @@
 #include "cuttlefish/host/libs/log_names/log_names.h"
 #include "cuttlefish/host/libs/metrics/device_metrics_orchestration.h"
 #include "cuttlefish/io/string.h"
+#include "cuttlefish/posix/remove.h"
 #include "cuttlefish/posix/symlink.h"
 #include "cuttlefish/process/command.h"
 #include "cuttlefish/result/result.h"
