+++ base/cvd/cuttlefish/host/commands/cvd/cli/log_files.cpp
@@ -31,10 +31,10 @@
 #include "absl/log/log.h"
 #include "absl/strings/match.h"
 
-#include "cuttlefish/common/libs/utils/files.h"
 #include "cuttlefish/files/directory_contents.h"
 #include "cuttlefish/files/directory_exists.h"
 #include "cuttlefish/host/commands/cvd/utils/common.h"
+#include "cuttlefish/posix/remove.h"
 #include "cuttlefish/result/result.h"
 
 namespace cuttlefish {
