+++ base/cvd/cuttlefish/host/commands/cvd/instances/BUILD.bazel
@@ -183,6 +183,7 @@ cf_cc_library(
         "//cuttlefish/host/libs/config:cuttlefish_config",
         "//cuttlefish/host/libs/metrics:device_metrics_orchestration",
         "//cuttlefish/host/libs/screen_recording_controls",
+        "//cuttlefish/posix:remove",
         "//cuttlefish/posix:symlink",
         "//cuttlefish/process:command",
         "//cuttlefish/process:managed_stdio",
@@ -247,6 +248,7 @@ cf_cc_library(
         "//cuttlefish/host/commands/cvd/instances:run_cvd_proc_collector",
         "//cuttlefish/host/commands/cvd/utils",
         "//cuttlefish/host/libs/config:config_constants",
+        "//cuttlefish/posix:remove",
         "//cuttlefish/posix:strerror",
         "//cuttlefish/process:command",
         "//cuttlefish/process:managed_stdio",
