+++ base/cvd/cuttlefish/host/commands/cvd/instances/lock/instance_lock.cpp
@@ -24,6 +24,7 @@
 #include <utility>
 
 #include "cuttlefish/common/libs/utils/files.h"
+#include "cuttlefish/posix/remove.h"
 #include "cuttlefish/posix/strerror.h"
 #include "cuttlefish/result/result.h"
