+++ base/cvd/cuttlefish/host/commands/run_cvd/BUILD.bazel
@@ -155,6 +155,7 @@ cf_cc_library(
         "//cuttlefish/host/libs/process_monitor",
         "//cuttlefish/host/libs/vm_manager",
         "//cuttlefish/io:string",
+        "//cuttlefish/posix:remove",
         "//cuttlefish/posix:strerror",
         "//cuttlefish/process:command",
         "//cuttlefish/process:execute",
