+++ base/cvd/cuttlefish/host/libs/command_util/snapshot_utils.cc
@@ -42,6 +42,7 @@
 #include "cuttlefish/host/libs/config/cuttlefish_config.h"
 #include "cuttlefish/posix/readlink.h"
 #include "cuttlefish/posix/realpath.h"
+#include "cuttlefish/posix/remove.h"
 #include "cuttlefish/posix/symlink.h"
 #include "cuttlefish/result/result.h"
