+++ base/cvd/cuttlefish/host/libs/web/cas/BUILD.bazel
@@ -16,6 +16,7 @@ cf_cc_library(
         "//cuttlefish/files:file_exists",
         "//cuttlefish/host/libs/web:android_build",
         "//cuttlefish/host/libs/web/cas:cas_flags",
+        "//cuttlefish/posix:remove",
         "//cuttlefish/process:command",
         "//cuttlefish/process:managed_stdio",
         "//cuttlefish/result",
@@ -34,6 +35,7 @@ cf_cc_test(
         "//cuttlefish/host/libs/web:android_build",
         "//cuttlefish/host/libs/web/cas:cas_downloader",
         "//cuttlefish/host/libs/web/cas:cas_flags",
+        "//cuttlefish/posix:remove",
         "//cuttlefish/result",
         "//cuttlefish/result:result_matchers",
         "//libbase",
